REFACTOR : Playlist Page에 들어가는 컴포넌트 분할
가독성 및 재사용을 위한 리팩토링
